Refs #292. migrate DetectorSystemFromXml
